Tweak install for our purposes.
Get latest OpenCV version 2.x.x.
Don't uninstall packages when handling dependencies.
Do install libav-tools, instead of ffmpeg.
Don't re-download the file if it already exists.
Modify the build flags.

+++ Ubuntu/opencv_install.sh
@@ -30,7 +30,9 @@ source dependencies.sh
 echo "--- Downloading OpenCV" $version
 mkdir -p $dldir
 cd $dldir
-wget -O $downloadfile http://sourceforge.net/projects/opencvlibrary/files/opencv-unix/$version/$downloadfile/download
+if [ ! -e $downloadfile ]; then
+  wget -O $downloadfile http://sourceforge.net/projects/opencvlibrary/files/opencv-unix/$version/$downloadfile/download
+fi
 
 echo "--- Installing OpenCV" $version
 echo $downloadfile | grep ".zip"
@@ -40,9 +42,12 @@ else
     tar -xvf $downloadfile
 fi
 cd opencv-$version
-mkdir build
+mkdir -p build
 cd build
-cmake -D CMAKE_BUILD_TYPE=RELEASE -D CMAKE_INSTALL_PREFIX=/usr/local -D WITH_TBB=ON -D BUILD_NEW_PYTHON_SUPPORT=ON -D WITH_V4L=ON -D INSTALL_C_EXAMPLES=ON -D INSTALL_PYTHON_EXAMPLES=ON -D BUILD_EXAMPLES=ON -D WITH_QT=ON -D WITH_OPENGL=ON ..
+# Use "-D CUDA_ARCH_BIN=3.2 -D CUDA_ARCH_PTX=3.2" like in 2.4/opencv2_4_10.sh?
+# What about "-D BUILD_TIFF=ON"?
+#cmake -D CMAKE_BUILD_TYPE=RELEASE -D CMAKE_INSTALL_PREFIX=/usr/local -D WITH_TBB=ON -D BUILD_NEW_PYTHON_SUPPORT=ON -D WITH_V4L=ON -D INSTALL_C_EXAMPLES=ON -D INSTALL_PYTHON_EXAMPLES=ON -D BUILD_EXAMPLES=ON -D WITH_QT=ON -D WITH_OPENGL=ON ..
+cmake -D CMAKE_BUILD_TYPE=RELEASE -D CMAKE_INSTALL_PREFIX=/usr/local -D WITH_TBB=ON -D BUILD_NEW_PYTHON_SUPPORT=ON -D WITH_V4L=ON -D INSTALL_C_EXAMPLES=ON -D INSTALL_PYTHON_EXAMPLES=ON -D BUILD_EXAMPLES=ON -D WITH_QT=ON -D WITH_OPENGL=ON -DWITH_OPENCL=OFF -DWITH_CUDA=ON ..
 make -j 4
 sudo make install
 sudo sh -c 'echo "/usr/local/lib" > /etc/ld.so.conf.d/opencv.conf'
